lang: Refactor serde => lang
